Revert "Merge branch 'development' into travis-ci"
This reverts commit c41e14a336ea69959683ea65530a0d9c5d61c444, reversing
changes made to b8dfc81d858f2e9b17410d8290489ec8903a2940.

+++ resources/user.py
@@ -1,6 +1,5 @@
 from flask_restful import Resource, reqparse
 from enum import Enum
-from resources.admin_required import admin_required
 from models.user import UserModel
 from models.revoked_tokens import RevokedTokensModel
 from werkzeug.security import safe_str_cmp
@@ -35,8 +34,13 @@ class UserRegister(Resource):
         return {"message": "User created successfully."}, 201
 
 class User(Resource):
-    @admin_required
+    @jwt_required
     def get(self, user_id):
+        #check if is_admin exist if not discontinue function
+        claims = get_jwt_claims()         
+        if not claims['is_admin']:
+            return {'message': 'Admin Access Required'}, 401
+
         user = UserModel.find_by_id(user_id)
 
         if not user:
@@ -52,36 +56,62 @@ class User(Resource):
             'archived': user.archived
         }, 200
         
-    @admin_required
+    @jwt_required
     def patch(self,user_id):
+        #check if is_admin exists - if not discontinue function
+        claims = get_jwt_claims()         
+        if not claims['is_admin']:
+            return {'message': "Admin Access Required"}, 401
+
         parser = reqparse.RequestParser()
         parser.add_argument('role',type=str,required=True,help="This field cannot be blank.")
 
         user = UserModel.find_by_id(user_id)
         if not user:
-            return {"Message": "Unable to update user"}, 400
+            return {"Message": "Unable to update user"}
 
         data = parser.parse_args()
         user.role = data['role']
         try:
             user.save_to_db()
         except:
-            return {'Message': 'An Error Has Occurred'}, 500
+            return {'Message': 'An Error Has Occured'}, 500
 
         return user.json(), 201
 
-    @admin_required
+    @jwt_required
     def delete(self, user_id):
+        #check if is_admin exists - if not discontinue function
+        claims = get_jwt_claims()         
+        if not claims['is_admin']:
+            return {'message': "Admin Access Required"}, 401
+
         user = UserModel.find_by_id(user_id)
         if not user:
-            return {"Message": "Unable to delete User"}, 400
+            return {"Message", "Unable to delete User"}, 404 
         user.delete_from_db()
         return {"Message": "User deleted"}, 200
 
+#pull all users - for debugging purposes disable before production
+class Users(Resource):
+    @jwt_required
+    def get(self):
+        #check if is_admin exists - if not discontinue function
+        claims = get_jwt_claims() 
+        if not claims['is_admin']:
+            return {'message': "Admin Access Required"}, 401
+        return {'Users': [user.json() for user in UserModel.query.all()]}
+
 class ArchiveUser(Resource):
 
-    @admin_required
+    @jwt_required
     def post(self, user_id):
+        #check if is_admin exist if not discontinue function
+        claims = get_jwt_claims() 
+
+        if not claims['is_admin']:
+            return {'message': "Admin Access Required"}, 401
+
         user = UserModel.find_by_id(user_id)
         if(not user):
             return{'Message': 'User cannot be archived'}, 400
@@ -127,8 +157,13 @@ class UsersRole(Resource):
     parser = reqparse.RequestParser()
     parser.add_argument('userrole',type=str,required=True,help="This field cannot be blank.")
 
-    @admin_required
+    @jwt_required
     def post(self):
+        #check if is_admin exist if not discontinue function
+        claims = get_jwt_claims() 
+        if not claims['is_admin']:
+            return {'message': "Admin Access Required"}, 401
+
         data = UsersRole.parser.parse_args()
         users = UserModel.find_by_role(data['userrole'])
         return {'users': [user.json() for user in users]}
